+++ classes/output/core_renderer.php
@@ -264,10 +264,27 @@ class core_renderer extends \theme_boost\output\core_renderer {
 
     public function get_frontdashboard_context() {
         global $CFG;
+
+        if (!empty($this->page->theme->settings->frontpagedashboardburgerbutton)) {
+            switch ($this->page->theme->settings->frontpagedashboardburgerbutton) {
+                case 1: // Show.
+                    $shownavdrawer = true;
+                    break;
+                case 2: // Hide.
+                    $shownavdrawer = false;
+                    break;
+                default: // Is 3 being 'capability'.
+                    $shownavdrawer = has_capability('theme/ned_boost:shownavdrawer', \context_course::instance($this->page->course->id)); // This should be the site id.
+                    break;
+            }
+        } else {
+            $shownavdrawer = has_capability('theme/ned_boost:shownavdrawer', \context_course::instance($this->page->course->id));
+        }
+
         require_once($CFG->libdir.'/behat/lib.php');
 
         $templatecontext = $this->get_dynamicbase();
-        $extraclasses = $this->get_navdraweropen($templatecontext, true);
+        $extraclasses = $this->get_navdraweropen($templatecontext, $shownavdrawer);
         $bodyattributes = $this->body_attributes($extraclasses);
         $templatecontext['bodyattributes'] = $bodyattributes;